make Alignment a Transformation
